requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
pydantic-settings
supabase
python-dotenv
requests
python-multipart

python-jose
PyJWT
passlib
cryptography

python-docx
PyPDF2
PyMuPDF
pdfplumber
pandas

pinecone
openai

langchain
langchain-openai
langchain-core
langchain-community

upstash-redis

aiosmtplib
apscheduler
storage3